Trait paperclip_openapi::v2::codegen::SchemaEmitter

pub trait SchemaEmitter {
type Definition: Schema + Debug;
    fn state(&self) -> &EmitterState;

    fn create_defs(&self, api: &Api<Self::Definition>) -> Result<(), Error> { ... }

    fn write_contents(&self, contents: String, path: &Path) -> Result<(), Error> { ... }

    fn generate_def_from_root(
        &self, 
        def: &Self::Definition
    ) -> Result<(), Error> { ... }

    fn build_def(
        &self, 
        def: &Self::Definition, 
        define: bool
    ) -> Result<String, Error> { ... }

    fn def_ns_name<'a>(
        &self, 
        def: &'a Self::Definition
    ) -> Result<Box<dyn Iterator<Item = String> + 'a>, Error> { ... }

    fn def_name(&self, def: &Self::Definition) -> Result<String, Error> { ... }

    fn def_mod_path(&self, def: &Self::Definition) -> Result<PathBuf, Error> { ... }

    fn emit_array(
        &self, 
        def: &Self::Definition, 
        define: bool
    ) -> Result<String, Error> { ... }

    fn emit_object(
        &self, 
        def: &Self::Definition, 
        define: bool
    ) -> Result<String, Error> { ... }

    fn try_emit_map(
        &self, 
        def: &Self::Definition, 
        define: bool
    ) -> Option<Result<String, Error>> { ... }

    fn emit_struct(&self, def: &Self::Definition) -> Result<String, Error> { ... }

    fn emit_type_alias(
        &self, 
        def: &Self::Definition, 
        ty: &str
    ) -> Result<String, Error> { ... }
}

SchemaEmitter is the first phase in codegen. It generates the relevant modules and API objects required for the next phase.

Associated Types

type Definition: Schema + Debug

The associated Schema implementor.

Loading content...
Required methods

fn state(&self) -> &EmitterState

Returns a reference to the underlying state.

Loading content...
Provided methods

fn create_defs(&self, api: &Api<Self::Definition>) -> Result<(), Error>

Entrypoint for emitter. Given an API spec, generate the definitions inside Rust modules in the configured working directory.

fn write_contents(&self, contents: String, path: &Path) -> Result<(), Error>

Writes the given contents to a file at the given path.

fn generate_def_from_root(&self, def: &Self::Definition) -> Result<(), Error>

Given a schema definition, generate the corresponding Rust definition.

NOTE: This doesn't generate any files. It only generates the code and writes it to EmitterState.

fn build_def(
&self,
def: &Self::Definition,
define: bool
) -> Result<String, Error>

Builds a given definition. Also takes a bool to specify whether we're planning to define a Rust type or whether we're reusing an existing type.

fn def_ns_name<'a>(
&self,
def: &'a Self::Definition
) -> Result<Box<dyn Iterator<Item = String> + 'a>, Error>

Returns an iterator of path components for the given definition.

NOTE: All components are snake_cased (including the definition name).

fn def_name(&self, def: &Self::Definition) -> Result<String, Error>

Returns the CamelCase name for the given definition.

fn def_mod_path(&self, def: &Self::Definition) -> Result<PathBuf, Error>

Returns the module path (from working directory) for the given definition.

fn emit_array(
&self,
def: &Self::Definition,
define: bool
) -> Result<String, Error>

Assumes that the given definition is an array and returns the corresponding vector type for it. Also takes a bool to specify whether we're defining the vector as a type alias or just fetching the type of the vector.

fn emit_object(
&self,
def: &Self::Definition,
define: bool
) -> Result<String, Error>

Assumes that the given definition is an object and returns the corresponding Rust struct. Also takes a bool to specify whether we're defining the struct or just reusing the type.

fn try_emit_map(
&self,
def: &Self::Definition,
define: bool
) -> Option<Result<String, Error>>

Checks if the given definition is a simple map and returns the corresponding BTreeMap. Also takes a bool to specify whether we're defining the map as a type alias or just fetching the type of the map.

fn emit_struct(&self, def: &Self::Definition) -> Result<String, Error>

Helper for emit_object - This returns the Rust struct definition for the given schema definition.

fn emit_type_alias(
&self,
def: &Self::Definition,
ty: &str
) -> Result<String, Error>

Given a definition and a type, emit the type alias using the definition's name.

Loading content...